Computation of Suitability Index (SI) and Reliability Index (RI)

The function allows to compute the Suitability Index (SI) and Reliability Index (RI) for a seed mixture or donor grassland composition to be used in a restoration site based on topographic features: elevation, slope and aspect.

see vignettes for additional details and examples

Usage

RestInd(trainingDB = NULL, composition, elevation, slope, aspect)

Arguments

trainingDB

dataframe with the species eligible for the statistical modelling, selected on the basis of their frequency and abundance in the vegetation and topographical variables database. Each species is associated with the values of topographic variables (elevation, slope and aspect) and its abundance. Default is NULL. When NULL, the training database is derived from 4081 vegetation surveys carried out in Piedmont Region, (North-Western Italy), which it includes 248 plant species eligible for modeling. A customized training database can be generated with trainingDB.

composition

dataframe with the seed mixture or donor grassland composition. It consists of two columns: 1) species code abbreviated in CEP names format and 2) abundance of the species. When "trainingDB" is NULL, the species CEP names are retrievable from cep.piem data. When "trainingDB" is not NULL, i.e. the user does not use default settings, species CEP names are retrievable from trainingDB, which generates CEP names of species suitable for modeling from a customized training database.

elevation

Elevation of restoration site, expressed in meters above sea level (m a.s.l.)

slope

Slope of restoration site, expressed in degrees (°)

aspect

Aspect of restoration site, expressed in degrees from North (°N)

Value

A list with three outputs:

DESCRIPTIVES: dataframe with the descriptive information related to modeled plant species:

cep.names

Species name in CEP format

species

Full species name

n.obs

Number of observations of each species selected for modeling

min.ele

Minimum elevation at which a species occurred

max.ele

Maximum elevation at which a species occurred

min.slope

Minimum slope at which a species occurred

max.slope

Maximum slope at which a species occurred

min.south

Minimum southness value at which a species occurred

max.south

Maximum southness value at which a species occurred

SPECIES ABUNDANCES: dataframe with the following information:

cep.names

Species name in CEP format

species

Full species name

PMA

Predicted Maximum Abundance. It is the maximum achievable abundance of a species in the restoration site, as predicted by the best model

POA

Predicted Optimal Abundance. It is the maximum achievable abundance of a species in its optimal ecological condition, based on all possible combinations of elevation, slope and southness.

ratio

Ratio between the PMA and POA. This ratio indicates how far (ratio = 0) or close (ratio = 1) a species is from its ecological optimum.

R2.adj

R square adjusted of the best Generalized Additive Model

RMSE

Root Mean Squared Error of the best Generalized Additive Model

SmDgA

Seed mixture or Donor grassland Abundance. Abundance of a species listed in the seed mixture or donor grassland composition imported by the user and eligible for modeling

EA

Expected Abundance. The highest achievable abundance of a species in a restoration site, based on how far the species is from the ecological optimum (i.e. computed from the multiplication of SmDgA by the ratio)

INDEXES:

SI

Suitability Index (SI). Suitability of a seed mixture or donor grassland to restore a site with specific topographic characteristics. It ranges between 0 and 1. When SI=0 the restoration site is totally beyond the optimal ecological ranges of all species of the seed mixture or donor grassland, which is therefore not appropriate for the site restoration. Conversely, when SI=1 the restoration site has the optimal ecological conditions for all species of the seed mixture or donor grassland, which is therefore perfectly appropriate for the site restoration.

RI

Reliability Index (RI). Index of the reliability of the Suitability Index (SI). The RI ranges between 0 and 1. When RI is close to 0 it means that few to none species contribute to the computation of the SI, whereas when RI is close to 1 the SI is computed with most to all species. Therefore, the higher is the RI, the most reliable is the SI. Not all the species of the seed mixture and donor grassland composition may modeled as i) they can be missing from the training database or ii) the values of the topographic factors of the restoration site are beyond their ecological ranges (e.g. if the elevation of the restoration site is 250 m and a species as an elevation range bounded between 1000 and 3000 m, such a species cannot be modeled).

Examples

#creation of seed mixture dataframe by retrieving species CEP names from ResNatSeed::cep.piem
mixture_composition<-data.frame(species=c("Festrubr","Chaehirs","Phleprat","Onobmont"),
                               sc=c(5,40,20,20))
#run function
ResNatSeed::RestInd(trainingDB = NULL,
                    composition = mixture_composition,
                    elevation = 1500,
                    aspect = 30,
                    slope = 5)
#> $DESCRIPTIVES
#>          cep.names                species n.obs min.ele max.ele min.slope
#> Chaehirs  Chaehirs Chaerophyllum hirsutum    43     623    2248         2
#> Festrubr  Festrubr          Festuca rubra    60     494    2743         2
#> Onobmont  Onobmont     Onobrychis montana    40    1395    2501      0.97
#> Phleprat  Phleprat        Phleum pratense    51     848    2439       2.6
#>          max.slope min.south max.south
#> Chaehirs        49       4.2       178
#> Festrubr        51      0.33       180
#> Onobmont        45       2.5       177
#> Phleprat        43         9       179
#> 
#> $SPECIES_ABUNDANCES
#>   cep.names                species  PMA  POA ratio R2.adj RMSE SmDgA   EA
#> 1  Chaehirs Chaerophyllum hirsutum 50.0 57.7  0.87   0.27 14.7    40 34.8
#> 2  Festrubr          Festuca rubra 90.4 96.4  0.94   0.89  8.5     5  4.7
#> 3  Onobmont     Onobrychis montana 20.2 85.6  0.24   0.57 15.3    20  4.8
#> 4  Phleprat        Phleum pratense  9.2 34.2  0.27   0.47  8.8    20  5.4
#> 
#> $INDEXES
#>     SI RI
#> 1 0.58  1
#>